refactor(step0): Station 패키지 구조 변경
