+++ src/main/java/nextstep/subway/station/domain/Station.java
@@ -1,4 +1,4 @@
-package nextstep.subway.station;
+package nextstep.subway.station.domain;
 
 import javax.persistence.*;
